setup.py: add repo option --repo-retry-fetches
Issue: LINUXEXEC-5665

After upgrade git-repo to v2.36.1, "repo sync" supports setting number
of times to retry fetches on transient errors, and the patches in
git-repo for extending retry machinism is dropped. Adding this repo
option to support retry fetches. Default retry times in git-repo is 2.
So this option accepts value >=3.

Set default value to 5 for this option to keep align with dropped
patches.

Signed-off-by: Changqing Li <changqing.li@windriver.com>