data/xml/meta-browser-dl.xml: update dl layer for meta-chromium
A tarball in meta-browser-dl-5.0 is very large (3.6G), which caused http
download failed on windshare.  In order to improve download experiences, we
force to clear up the big tarball in meta-browser-dl-5.0, and split it into 2
pieces, and save them to meta-browser-dl-5.0.part01 and
meta-browser-dl-5.0.part02.

The layer meta-browser-dl-5.0 provides a post hook to merge the fragment into a
tarball for setup.  Place sub repo as sub dir of meta-browser-dl-5.0, map to
layers/meta-browser-dl/meta-browser-dl.part01 and
layers/meta-browser-dl/meta-browser-dl.part02

According to the experience in download testing, size of git repo <2GB is not
likely to be broken, so split 2 parts is a decent start and we can consider
3,4,5, ... if someone has trouble with the 1.8 GB files in future.

Issue: LINUXEXEC-32713

Signed-off-by: Hongxu Jia <hongxu.jia@windriver.com>