bin/setup.py: Don't set revision if it is the same as default
Issue: LIN1024-3038

* Set revision for each repo when it is same as default revision doesn't make
sense.

* Fixed:
$ ./wrlinux-x/setup.sh --machine qemux86-64 --layers meta-intel-qat-dl

WARNING: Found duplicated basename in default.xml's attribute 'path':
WARNING: layers/meta-intel-qat-dl

This was because there is a 'path="layers/meta-intel-qat-dl"' in
data/xml/meta-intel-qat-dl.xml, it will be pulled in directly, but the one
specified by --layers will be added with an extra '</project>' which makes the
remove duplicates doesn't work, this patch also fixed the problem.

Signed-off-by: Robert Yang <liezhi.yang@windriver.com>